make zephyr-vm serialization compatible

+++ charming/src/component/angle_axis.rs
@@ -18,11 +18,11 @@ pub struct AngleAxis {
 
     /// The index of angle axis in Polar Coordinate.
     #[serde(skip_serializing_if = "Option::is_none")]
-    polar_index: Option<f64>,
+    polar_index: Option<i64>,
 
     /// Starting angle of axis, default to 90.
     #[serde(skip_serializing_if = "Option::is_none")]
-    start_angle: Option<f64>,
+    start_angle: Option<i64>,
 
     /// Whether the direction of axis is clockwise, default to true.
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -34,36 +34,36 @@ pub struct AngleAxis {
     type_: Option<AxisType>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    zlevel: Option<f64>,
+    zlevel: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    z: Option<f64>,
+    z: Option<i64>,
 
     /// The minimun value of axis.
     #[serde(skip_serializing_if = "Option::is_none")]
-    min: Option<f64>,
+    min: Option<i64>,
 
     /// The maximum value of axis.
     #[serde(skip_serializing_if = "Option::is_none")]
-    max: Option<f64>,
+    max: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     scale: Option<bool>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    split_number: Option<f64>,
+    split_number: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    min_interval: Option<f64>,
+    min_interval: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    max_interval: Option<f64>,
+    max_interval: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    interval: Option<f64>,
+    interval: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    log_base: Option<f64>,
+    log_base: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     silent: Option<bool>,
@@ -142,12 +142,12 @@ impl AngleAxis {
         self
     }
 
-    pub fn polar_index<F: Into<f64>>(mut self, polar_index: F) -> Self {
+    pub fn polar_index<F: Into<i64>>(mut self, polar_index: F) -> Self {
         self.polar_index = Some(polar_index.into());
         self
     }
 
-    pub fn start_angle<F: Into<f64>>(mut self, start_angle: F) -> Self {
+    pub fn start_angle<F: Into<i64>>(mut self, start_angle: F) -> Self {
         self.start_angle = Some(start_angle.into());
         self
     }
@@ -162,22 +162,22 @@ impl AngleAxis {
         self
     }
 
-    pub fn zlevel<F: Into<f64>>(mut self, zlevel: F) -> Self {
+    pub fn zlevel<F: Into<i64>>(mut self, zlevel: F) -> Self {
         self.zlevel = Some(zlevel.into());
         self
     }
 
-    pub fn z<F: Into<f64>>(mut self, z: F) -> Self {
+    pub fn z<F: Into<i64>>(mut self, z: F) -> Self {
         self.z = Some(z.into());
         self
     }
 
-    pub fn min<F: Into<f64>>(mut self, min: F) -> Self {
+    pub fn min<F: Into<i64>>(mut self, min: F) -> Self {
         self.min = Some(min.into());
         self
     }
 
-    pub fn max<F: Into<f64>>(mut self, max: F) -> Self {
+    pub fn max<F: Into<i64>>(mut self, max: F) -> Self {
         self.max = Some(max.into());
         self
     }
@@ -187,27 +187,27 @@ impl AngleAxis {
         self
     }
 
-    pub fn split_number<F: Into<f64>>(mut self, split_number: F) -> Self {
+    pub fn split_number<F: Into<i64>>(mut self, split_number: F) -> Self {
         self.split_number = Some(split_number.into());
         self
     }
 
-    pub fn min_interval<F: Into<f64>>(mut self, min_interval: F) -> Self {
+    pub fn min_interval<F: Into<i64>>(mut self, min_interval: F) -> Self {
         self.min_interval = Some(min_interval.into());
         self
     }
 
-    pub fn max_interval<F: Into<f64>>(mut self, max_interval: F) -> Self {
+    pub fn max_interval<F: Into<i64>>(mut self, max_interval: F) -> Self {
         self.max_interval = Some(max_interval.into());
         self
     }
 
-    pub fn interval<F: Into<f64>>(mut self, interval: F) -> Self {
+    pub fn interval<F: Into<i64>>(mut self, interval: F) -> Self {
         self.interval = Some(interval.into());
         self
     }
 
-    pub fn log_base<F: Into<f64>>(mut self, log_base: F) -> Self {
+    pub fn log_base<F: Into<i64>>(mut self, log_base: F) -> Self {
         self.log_base = Some(log_base.into());
         self
     }
